refactor(host_runtime): extract trigger_management tests to sibling file
Thermo review flagged trigger_management.rs at 1268 lines (grew past
the file-size guideline while absorbing #5886's active_hold tests).
Move the inline #[cfg(test)] module to trigger_management/tests.rs,
matching the facade.rs/facade/tests.rs pattern already used elsewhere
in this codebase. Pure move, no behavior or test-logic change.
